+++ lib/queue/index.ts
@@ -5,3 +5,9 @@ export {
   enqueueReceiptProcessing,
 } from "./receipt-queue";
 export type { ReceiptProcessingJobData } from "./receipt-queue";
+export {
+  REVIEW_DISABLE_QUEUE,
+  getReviewDisableQueue,
+  enqueueReviewDisable,
+} from "./review-disable-queue";
+export type { ReviewDisableJobData } from "./review-disable-queue";